server aula 3 rotas autenticadas

+++ NLW-Spacetime/server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-	app.get("/memories", async () => {
+	// vai executar antes de todas as rotas para verificar se o usuário está logado
+	app.addHook("preHandler", async (req) => {
+		// jwtVerify() verifica se a requisição para está rota possui token de autenticação
+		await req.jwtVerify();
+	});
+
+	app.get("/memories", async (req) => {
 		const memories = await prisma.memory.findMany({
+			where: {
+				userId: req.user.sub, // vai trazer somente as memorias do usuário logado
+			},
 			orderBy: {
 				createdAt: "asc",
 			},
@@ -19,7 +28,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		});
 	});
 
-	app.get("/memories/:id", async (req) => {
+	app.get("/memories/:id", async (req, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -30,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
 				id,
 			},
 		});
+
+		if (!memory.isPublic && memory.userId !== req.user.sub) {
+			return reply.status(401).send();
+		}
 
 		return memory;
 	});
@@ -48,14 +61,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
 				content,
 				coverUrl,
 				isPublic,
-				userId: "05a522fb-97d5-4e53-bd03-bcd52afb69bc",
+				userId: req.user.sub,
 			},
 		});
 
 		return memory;
 	});
 
-	app.put("/memories", async (req) => {
+	app.put("/memories", async (req, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -70,7 +83,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
 		const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-		const memory = await prisma.memory.update({
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
+
+		if (memory.userId !== req.user.sub) {
+			return reply.status(401).send();
+		}
+
+		memory = await prisma.memory.update({
 			where: { id },
 			data: {
 				content,
@@ -82,11 +105,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		return memory;
 	});
 
-	app.delete("/memories", async (req) => {
+	app.delete("/memories", async (req, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
+
 		const { id } = paramsSchema.parse(req.params);
+
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
+
+		if (memory.userId !== req.user.sub) {
+			return reply.status(401).send();
+		}
 
 		await prisma.memory.delete({
 			where: {
